requirements.txt
fastapi
uvicorn
prometheus-client
pydantic
pandas
pyarrow
confluent-kafka
python-dotenv
scikit-learn
numpy
pytest
pandas
numpy
scikit-learn
scipy
boto3
redis


